Refactor the writer interface and create buffered writers
Make Writer into an interface that exposes `io.Writer` and the protocol
that should be written to the underlying writer.

Include the precision within the protocol itself too.

Read global error messages from the result.

Include a BufferedWriter that ensures that different lines of the
protocol do not get split. WritePoints will now call Write once for
every line of the line protocol and the BufferedWriter should be used to
control when it writes. WritePoints will automatically call Flush on the
Writer that's passed in if it has that function, otherwise it will
assume everything was written directly to the underlying io.Writer.

Also includes a TimedWriter. The TimedWriter wraps the BufferedWriter
and will automatically flush the BufferedWriter at a duration. The
TimedWriter must be stopped after it is created to avoid leaking a
goroutine.

Modifies WritePoints to flush the buffer of the passed in Writer
automatically if it exists. This feature is exclusive to WritePoints and
does not affect `(*Point).WriteTo`.

Adds a buffer pool for writing points to the v1 line protocol. This way,
you cannot accidentally call `Encode()` and get a half-written point to
the `io.Writer`.